requirements.txt
# Python 3.13.3
boto3==1.38.36
genanki==0.13.1
python-dotenv==1.1.0
tqdm==4.67.1
